Add HTTP client pseudo-node
* Refactors `oak_runtime::node::http` to have separate sub-modules for
  `client` and `server`.
* Adds `HttpClientNode`, with support for both `HTTP` and `HTTPS` URI schemes.
* New tests are added to test the HTTP client functionality.
* Replaces `oak_runtime::node::grpc::invocation` with a generic
`oak_runtime::node::invocation` for both gRPC and HTTP client nodes.
