+++ examples/deny.toml
@@ -61,6 +61,10 @@ version = "=0.10.2"
 name = "parking_lot_core"
 version = "=0.7.2"
 
+[[bans.skip]]
+name = "webpki-roots"
+version = "=0.19.0"
+
 # List of allowed licenses.
 # For more detailed information see http://go/thirdpartylicenses.
 [licenses]
