+++ experimental/deny.toml
@@ -32,6 +32,10 @@ version = "=0.4.27"
 name = "cfg-if"
 version = "=0.1.10"
 
+[[bans.skip]]
+name = "webpki-roots"
+version = "=0.19.0"
+
 # List of allowed licenses.
 # For more detailed information see http://go/thirdpartylicenses.
 [licenses]
