+++ oak_runtime/src/node/http/tests.rs
@@ -14,21 +14,35 @@
 // limitations under the License.
 //
 
-use super::*;
 use crate::{
-    io::{ReceiverExt, SenderExt},
-    node::{Node, NodeIsolation},
-    proto::oak::invocation::HttpInvocation,
+    io::{ReceiverExt, Sender, SenderExt},
+    node::Node,
+    proto::oak::invocation::{HttpInvocation, HttpInvocationSender},
+    RuntimeProxy,
 };
+use log::{error, info};
+use oak_io::{
+    handle::{ReadHandle, WriteHandle},
+    Receiver,
+};
+use oak_services::proto::oak::encap::{HttpRequest, HttpResponse};
+use std::io;
+use tokio::sync::oneshot;
+
+use crate::node::NodeIsolation;
+use std::sync::mpsc;
+
 use maplit::hashmap;
 use oak_abi::{
     label::{confidentiality_label, public_key_identity_tag, tls_endpoint_tag, Label},
     proto::oak::application::{
-        node_configuration::ConfigType, ApplicationConfiguration, HttpServerConfiguration,
-        NodeConfiguration,
+        node_configuration::ConfigType, ApplicationConfiguration, HttpClientConfiguration,
+        HttpServerConfiguration, NodeConfiguration,
     },
+    OakStatus,
 };
-use oak_io::Sender;
+use oak_io::OakError;
+use prost::Message;
 use std::fs;
 
 const LOCAL_CA: &str = "../examples/certs/local/ca.pem";
@@ -87,7 +101,7 @@ impl Node for RouterNode {
             runtime
                 .node_register(
                     Box::new(echo_node),
-                    "echo_name",
+                    "echo_node",
                     &request_label,
                     read_handle,
                 )
@@ -361,7 +375,13 @@ fn create_runtime() -> RuntimeProxy {
     .expect("Could not create TLS config from local certs.");
     let secure_server_config = crate::SecureServerConfiguration {
         grpc_config: None,
-        http_config: Some(crate::HttpConfiguration { tls_config }),
+        http_config: Some(crate::HttpConfiguration {
+            tls_config,
+            http_client_root_tls_certificate: include_bytes!(
+                "../../../../examples/certs/local/ca.pem"
+            )
+            .to_vec(),
+        }),
     };
     let signature_table = crate::SignatureTable::default();
     info!("Create runtime for test");
@@ -391,7 +411,7 @@ fn create_server_node(runtime: &RuntimeProxy, port: u32) -> oak_abi::Handle {
 fn create_communication_channel(runtime: &RuntimeProxy) -> (oak_abi::Handle, oak_abi::Handle) {
     // create channel: one end to server_node::run; the other to the Oak node.
     let (init_sender, init_receiver) = runtime
-        .channel_create("HTTP init", &Label::public_untrusted())
+        .channel_create("HTTP server init", &Label::public_untrusted())
         .expect("Could not create channel");
 
     // At the start the HTTP server pseudo-Node expects to receive an invocation channel, with
@@ -399,7 +419,7 @@ fn create_communication_channel(runtime: &RuntimeProxy) -> (oak_abi::Handle, oak
     //
     // Create a channel for receiving invocations to pass to the HTTP server pseudo-Node.
     let (invocation_sender, invocation_receiver) = runtime
-        .channel_create("HTTP invocation", &Label::public_untrusted())
+        .channel_create("HTTP server invocation", &Label::public_untrusted())
         .expect("Could not create channel");
     let invocation_sender = HttpInvocationSender {
         sender: Some(Sender::<HttpInvocation>::new(WriteHandle {
@@ -504,5 +524,228 @@ async fn send_request(
                 }
             }
         }
+    }
+}
+
+#[test]
+fn test_https_client_can_handle_https_requests_to_an_external_service() {
+    init_logger();
+
+    let runtime = create_runtime();
+
+    // Create an HTTP client pseudo-node to serve requests to `https://www.google.com/`. A valid
+    // authority is expected in the URI, so we cannot instead start a server on localhost and use
+    // that as the external service. The HTTP client pseudo-node is created with a communication
+    // channel. A handle to the read half of the channel is returned. The Oak node will use it
+    // to fetch the write handle to an invocation channel for sending the invocations to the
+    // HTTP client pseudo-node.
+    let oak_node_init_receiver = create_client_node(&runtime, "www.google.com".to_string());
+
+    // Create a sync_channel to be notified when the Oak Node is completed.
+    let (result_sender, result_receiver) = mpsc::sync_channel(1);
+
+    // Create a test Oak node that sends requests to the HTTP client pseudo-node.
+    let client_test_node = ClientTesterNode {
+        uri: "https://www.google.com/".to_string(),
+        result_sender,
+    };
+
+    // Register the test Oak node in the runtime.
+    runtime
+        .node_register(
+            Box::new(client_test_node),
+            "client_test_node",
+            &Label::public_untrusted(),
+            oak_node_init_receiver,
+        )
+        .unwrap();
+
+    // Wait for the test Node to complete execution before terminating the Runtime.
+    let result_value = result_receiver.recv().expect("test node disconnected");
+
+    assert!(result_value.is_ok());
+    let resp = result_value.unwrap();
+    // The test uses a self signed certificate. So, we expect an error that is converted to an
+    // INTERNAL_SERVER_ERROR response in the HTTP client node.
+    assert_eq!(
+        resp.status,
+        http::StatusCode::INTERNAL_SERVER_ERROR.as_u16() as i32
+    );
+    runtime.runtime.stop();
+}
+
+#[test]
+fn test_https_client_can_handle_http_requests_to_an_external_service() {
+    init_logger();
+
+    let runtime = create_runtime();
+    // This is a public node, so the authority can be empty.
+    let empty_authority = "".to_string();
+    // Create an HTTP client pseudo-node to serve `HTTP` requests`.
+    let oak_node_init_receiver = create_client_node(&runtime, empty_authority);
+
+    // Create a sync_channel to be notified when the Oak Node is completed.
+    let (result_sender, result_receiver) = mpsc::sync_channel(1);
+
+    // Create a test Oak node that sends requests to the HTTP client pseudo-node.
+    let client_test_node = ClientTesterNode {
+        uri: "http://www.google.com".to_string(),
+        result_sender,
+    };
+
+    // Register the test Oak node in the runtime.
+    runtime
+        .node_register(
+            Box::new(client_test_node),
+            "client_test_node",
+            &Label::public_untrusted(),
+            oak_node_init_receiver,
+        )
+        .unwrap();
+
+    // Wait for the test Node to complete execution before terminating the Runtime.
+    let result_value = result_receiver.recv().expect("test node disconnected");
+
+    assert!(result_value.is_ok());
+    let resp = result_value.unwrap();
+    assert_eq!(resp.status, http::StatusCode::OK.as_u16() as i32);
+    runtime.runtime.stop();
+}
+
+/// Creates an HTTP client pseudo-node in the given Runtime.
+fn create_client_node(runtime: &RuntimeProxy, authority: String) -> oak_abi::Handle {
+    let (init_receiver, invocation_receiver) = create_http_client_communication_channel(runtime);
+    let client_config = NodeConfiguration {
+        config_type: Some(ConfigType::HttpClientConfig(HttpClientConfiguration {
+            authority,
+        })),
+    };
+
+    runtime
+        .node_create(
+            "test_http_client",
+            &client_config,
+            &Label::public_untrusted(),
+            invocation_receiver,
+        )
+        .expect("Could not create HTTP client node!");
+
+    init_receiver
+}
+
+/// An HTTP client pseudo-node needs a channel to read `HttpInvocation`s from.
+/// We wrap the write half of the channel in an `HttpInvocationSender` and hand it to the Oak node,
+/// which will use the write half of the channel to send the requests to the HTTP client
+/// pseudo-node.
+/// This function creates two channels, and returns their receiver (read) ends. The first receiver
+/// will be used as the initial handle to the Oak node. The second receiver is the initial handle
+/// to the HTTP client node.
+fn create_http_client_communication_channel(
+    runtime: &RuntimeProxy,
+) -> (oak_abi::Handle, oak_abi::Handle) {
+    // create HttpInvocation channel: The receiver end goes to the Oak node. The other end goes to
+    // the HTTP client pseudo-node.
+    let (init_sender, init_receiver) = runtime
+        .channel_create("Oak node init", &Label::public_untrusted())
+        .expect("Could not create channel");
+
+    // create HttpInvocationSender channel: At the start, the Oak Node expects to receive an
+    // invocation channel, with exactly one handle in it.
+    //
+    // Create a channel for sending invocations to the HTTP client pseudo-Node.
+    let (invocation_sender, invocation_receiver) = runtime
+        .channel_create("HTTP client invocation", &Label::public_untrusted())
+        .expect("Could not create channel");
+    let invocation_sender = HttpInvocationSender {
+        sender: Some(Sender::<HttpInvocation>::new(WriteHandle {
+            handle: invocation_sender,
+        })),
+    };
+    let init_sender = Sender::<HttpInvocationSender>::new(WriteHandle {
+        handle: init_sender,
+    });
+
+    if let Err(error) = init_sender.send(invocation_sender, runtime) {
+        panic!("Could not write to the `init_sender` channel: {}", error);
+    }
+    if let Err(error) = init_sender.close(runtime) {
+        panic!("Could not close the `init_sender` channel: {}", error);
+    }
+
+    (init_receiver, invocation_receiver)
+}
+
+struct ClientTesterNode {
+    uri: String,
+    result_sender: mpsc::SyncSender<Result<HttpResponse, OakError>>,
+}
+
+impl Node for ClientTesterNode {
+    fn node_type(&self) -> &'static str {
+        "client-tester"
+    }
+    fn isolation(&self) -> NodeIsolation {
+        // Even though this node is not actually sandboxed, we are simulating a Wasm node during
+        // testing.
+        NodeIsolation::Sandboxed
+    }
+    fn run(
+        self: Box<Self>,
+        runtime: RuntimeProxy,
+        handle: oak_abi::Handle,
+        _notify_receiver: oneshot::Receiver<()>,
+    ) {
+        // At start-of-day we need/expect to receive a write handle for an invocation channel
+        // to use for all subsequent activity.
+        let startup_receiver = Receiver::<HttpInvocationSender>::new(ReadHandle { handle });
+        let invocation_sender = startup_receiver
+            .receive(&runtime)
+            .and_then(|invocation_sender| {
+                invocation_sender
+                    .sender
+                    .ok_or(OakError::OakStatus(OakStatus::ErrBadHandle))
+            })
+            .expect("Failed to retrieve invocation channel write handle");
+        if let Err(err) = startup_receiver.close(&runtime) {
+            error!(
+                "Failed to close initial inbound channel {}: {:?}",
+                handle, err
+            );
+        }
+
+        // create request.
+        // NOTE: Method is case sensitive.
+        let request = HttpRequest {
+            uri: self.uri.clone(),
+            method: http::Method::GET.to_string(),
+            body: vec![],
+            headers: None,
+        };
+        // create channel
+        let pipe = crate::node::http::util::Pipe::new(
+            &runtime,
+            &Label::public_untrusted(),
+            &Label::public_untrusted(),
+        )
+        .expect("Could not create the Pipe");
+
+        // send the request on invocation_sender
+        pipe.insert_message(&runtime, request)
+            .expect("Could not insert HTTP request in the pipe");
+
+        // send the invocation to the HTTP client pseudo-node
+        pipe.send_invocation(&runtime, invocation_sender.handle.handle)
+            .expect("Could not send the invocation");
+
+        // wait for the response to come
+        let response_receiver = crate::io::Receiver::<HttpResponse>::new(ReadHandle {
+            handle: pipe.response_reader,
+        });
+        let response = response_receiver.receive(&runtime);
+
+        // notify the test
+        self.result_sender
+            .send(response)
+            .expect("could not send the result");
     }
 }
