+++ sdk/deny.toml
@@ -35,6 +35,10 @@ version = "=0.4.22"
 name = "cfg-if"
 version = "=0.1.10"
 
+[[bans.skip]]
+name = "webpki-roots"
+version = "=0.19.0"
+
 # List of allowed licenses.
 # For more detailed information see http://go/thirdpartylicenses.
 [licenses]
